Fix image indexing
