+++ .storybook/compile.ts
@@ -14,10 +14,7 @@ const prepareStory = (story: StoryConfig) => {
 
 export interface CompileOptions {}
 
-export const compile = async (
-  config: DynamicConfig,
-  options?: CompileOptions
-) => {
+export const compile = async (config: DynamicConfig, options?: CompileOptions) => {
   const { baseCsf } = config;
   const stories = await config.stories();
