--- components/dynamic/index.ts
@@ -1,3 +1,0 @@
-import type { DynamicConfig } from './types';
-export const defineStories = async (config: DynamicConfig) => config;
-export * from './types';
