--- components/fixtures/stories.json
@@ -1,12 +1,0 @@
-{
-  "states": [
-    {
-      "name": "Badge",
-      "filePath": "Badge-Basic.png"
-    },
-    {
-      "name": "Button",
-      "filePath": "MyButton-Basic.png"
-    }
-  ]
-}
